requirements.txt
Django==2.2.6
djangorestframework==3.10.3
mysqlclient==1.4.4
numpy==1.17.3
pandas==0.25.2
ply==3.11
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
xlrd==1.2.0
